Refactor "Dialog" Services
